data/pyproject.toml
[project]
name = "study_companion"
requires-python = ">=3.10,<3.15"
version = "0.1.0"
dependencies = [
    "dagster==1.12.11",
    "dagster-postgres>=0.28.11",
    "dagster-openai>=0.25.0",
    "pypdf>=5.0.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.3.0",
    "requests>=2.31.0",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/study_companion"]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "study_companion"
registry_modules = [
    "study_companion.components.*",
]


